bmi055 split accel and gyro headers
